Fix TouchEventAck handling by routing it to root view.
This CL tentatively fixes a crash (b/349075021) by correcting the
handling of TouchEventAcks in the TouchEventQueue. Previously,
TouchEventAcks were sent directly to the |target_view|, even if it
wasn't the root view. This change ensures that TouchEvent acking is
always routed through the root view for proper processing. This CL also
adds a check to confirm that the view is guaranteed to be a top-level
view when we don't have an input event router.

Bug: b:349075021
Change-Id: I376ada8c82e3871786d158dba89e5a79ed72e243
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5664654
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Aman Verma <amanvr@google.com>
Cr-Commit-Position: refs/heads/main@{#1320651}