[RasterInducingScroll] Collect discardable images under DrawScrollingContentsOp
It's like how we rasterize a DrawScrollingContentsOp with the scroll
offset applied and collect discardable images from the display item
list of the DrawScrollingContentsOp recursively.

As the discardable image map can change on change of scroll offsets,
without a change of the RasterSource or the DisplayItemList, and the
pending tree and the active tree may have different current scroll
offsets, the discardable image map is moved from DisplayItemList to
PictureLayerImpl. It's re-generated in two cases:
1. During commit, when we have a new DisplayItemList;
2. During raster-inducing scroll impl-side invalidation, when any
   raster-inducing scroll offsets changed.

During activation, the pending tree discardable image map is pushed to
the active tree. If the map is different, we'll re-register animation
images in the active tree.

Bug: 40517276
Change-Id: I84a629259e9bae7b39c4d74169448e0005b8bfde
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5635849
Commit-Queue: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320673}