[PDF Ink Signatures] Expose brush get color for production
When saving a stroke to a PDF it is necessary to know the color to draw.
Change the existing InkBrush::GetColorForTesting() API to be just
GetColor(), so it can be used in production code for saving.

Bug: 335517469
Change-Id: I0197699e8bfe9ee50d7a1292953ad3b819053566
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5656114
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Alan Screen <awscreen@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1320684}