[VCN] Add CVC chip button to the manual fill view
Add a CVC labeled chip button to both original manual fill and the
upgraded manual fill view. This button is only shown for virtual cards
for now (in the future this may be shown for cards that have CVC
stored - not yet supported on Bling). The text value will be an
obfuscated 3-digit *** string for regular cards or 4 digit **** string
for Amex cards.

Bug: 347325672
Change-Id: Iab3f7ced54a2afbb34b611e29952c99aafbb125e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5649111
Reviewed-by: Noémie St-Onge <noemies@google.com>
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Siyu An <siyua@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320708}