CrOS WebUI: Use assert helpers from chai_assert.js instead of window.
This is in preparation of rolling Chai to v5, which breaks js2gtest
usage of Chai, since in v5 it is now only available as an ES Module and
not as a global 'window.chai' object.

Updating some CrOS usages in an effort to avoid having to keep Chai v4
(side-by-side with v5) for remaining js2gtest users.

Bug: 347991521
Change-Id: I52f5e4be9ceb58037b91411ca34be79bf8f8f7b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5665054
Auto-Submit: Demetrios Papadopoulos <dpapad@chromium.org>
Commit-Queue: Wes Okuhara <wesokuhara@google.com>
Reviewed-by: Wes Okuhara <wesokuhara@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1320728}