[iOS] Add strings to present passkeys in the CPE
This CL adds strings to present passkey information in the CPE and
includes a minimal change to include each new string in the code.

It also renames a few variable and methods using the word "password"
so that they now use the word "credential" instead, when they apply
to both passwords and passkeys.

Bug: 330355124
Change-Id: I5810e85cc6d34390fc6eb90ee65f7a8f231313cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5661377
Reviewed-by: Tommy Martino <tmartino@chromium.org>
Commit-Queue: Alexis Hétu <sugoi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320762}