Ensure base::File checksum_file is closed on task_runner_
When an error happens, any open base::File which is either a parameter,
or is attached to the unused callback is destroyed, and the close
operation is blocking which causes chrome to crash.

The base::File or callback with one needs to be moved into the
task_runner_ to be destroyed. This stores checksum_file and
firmware_file locally and ensures cleanup is always done on
task_runner_.

Change-Id: I8ef386ebd69a4f4e29566d8d551b4868ae598719
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5650513
Auto-Submit: Joel Hockey <joelhockey@chromium.org>
Commit-Queue: Joel Hockey <joelhockey@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320786}