Fix ServiceWorkersDoNotTimeOutWithPolicy test and enable on all platforms
This commit makes the following changes in `ServiceWorkerLifetimeKeepaliveBrowsertest`

1. The `ExtendedBackgroundLifetimeForPortConnectionsToUrls` policy is
supported on all platforms. However, the tests were only built for ChromeOS.
This change removes the buildflag check for the policy tests.

2. The `ServiceWorkersDoNotTimeOutWithPolicy` test had a cleanup logic to
stop the service workers before browser shutdown. It looks like the logic
was added to avoid hitting a DCHECK at shutdown. This should no longer be
needed after crrev.com/c/5202906.

3. The `ServiceWorkersDoNotTimeOutWithPolicy` test was flaky. The test
loads two extensions, waits for them to establish a port connection,
advances the clock and expects the service workers to be alive. If the
clock is advanced before the service worker is activated (and the
activation request is in-flight before advancement), the activation
request will expire, leading to a termination of the service worker
(because activation requests have KILL_ON_TIMEOUT behavior) and fail the
test. Thus, we need to ensure that the clock is advanced after the worker
is activated.

Bug: 40272276,41494602
Change-Id: I69b40c1ddaebd31ea555bf070d6daadc6ae1f4d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5662060
Reviewed-by: Emilia Paz <emiliapaz@chromium.org>
Commit-Queue: Sohail Rajdev <sorajdev@microsoft.com>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320792}