Close Picker submenu if user resumes keyboard navigation in PickerView.
This is an edge case where the user can open a submenu with mouse hover
while they are using keyboard to navigate the main PickerView. Since the
keyboard selection can be separate to the mouse hover selection (similar
to omnibox behaviour), we just close the submenu if the user resumes
keyboard navigation in the main PickerView.

Bug: b:349259687
Change-Id: I3e320750cebd4d0a8ff34488416978197bd2d881
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5665515
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Michelle Chen <michellegc@google.com>
Cr-Commit-Position: refs/heads/main@{#1320796}