media/gpu/vaapi: USe ScopedID<VASurfaceID> ISO VASurface
VASurface is used for Vaapi*Picture classes (used for encoding and
decoding) and also for VaapiDecodeSurfaceHandler's implementation.

Decoding: ref-counted VASurfaces are unnecessary since Vaapi*Pictures
are ref-counted themselves (that's needed for Duplicate() calls) and
libva's VASurfaces don't need to be destroyed (they are owned by
VaapiVideoDecoder::allocated_va_surfaces_); the only needed
functionality is a release callback to VaapiVD::ReleaseVideoFrame.

Encoding: same rationale but VaapiVEA does not keep ownership of any
allocated libva VASurfaces; this CL sorts that out by moving the
ScopedVASurface into the release callback, where it's bound until
this is called.

It's therefore simpler to just use ScopedID<VASurfaceID>, which holds
an integer type and a release callback. This CL introduces that and
calls it VASurfaceHandle.

FTR this passes:
- video.Play.(av1|vp9)_hw_*
- video.ChromeStackDecoder.(vp9|av1)*
- video.EncodeAccel.* video.WebCodecs.* webrtc.RTCPeerConnection.*
- video.Seek.av1 video.Seek.vp9

on rex.

Bug: 339518553
Change-Id: Id2523c7dc787122c6d6348f795c17ce1b9e4d2bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5590385
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320803}