ozone/drm: use *_NON_SIMPLE_OVERLAYS_FAILED if fullscreen overlay fails
This CL prepares ozone/drm to use
gfx::SwapResult::SWAP_NON_SIMPLE_OVERLAYS_FAILED if
|notify_fullscreen_overlay_failure_| is set.

1) |notify_fullscreen_overlay_failure_| will be set by the owner
of the Presenter based on the kHandleOverlaysSwapFailure.
Ozone/Drm can't check that feature directly as it creates
a circular dependency. A CL that does so will be sent later once
all the necessary bits land.
2) SWAP_NON_SIMPLE_OVERLAYS_FAILED is used only for fullscreen
overlays. In the future, it can also cover other types of overlays
if needed.
3) gfx::SwapResult::SWAP_FAILED is changed to
::SWAP_NON_SIMPLE_OVERLAYS_FAILED for fullscreen overlays. However,
the gpu watchdog is still triggered. It'll be reset once the
frame is rescheduled (happens as soon as viz gets a swap failure
event) or a new frame is sent (if has already been queued).

Bug: 331237773
Change-Id: I3ed389f027b9beb4b1e6804ce03cf88e1ea2154d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5630560
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Reviewed-by: Maggie Chen <magchen@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1320830}