ModelTypeSyncBridge: Improve async Get[All]Data default impls
GetDataForCommit() and GetAllDataForDebugging() are being migrated from
asynchronous (returning the result via a callback) to synchronous
(returning the result directly).

For now (as of crrev.com/c/5659558), the ModelTypeSyncBridge
interface contains both, with the old asynchronous versions by
default calling into the new synchronous ones.
Previously, the default impls unconditionally ran their callback - in
case of errors, with a null result. That was a behavior change from the
traditional behavior, which was to not run the callback at all in that
case.
To avoid accidental behavior changes during the transition, this CL
updates the default impls to only run the callback with a non-null
result.

Bug: 347649753
Change-Id: I6f920498dd7c890033d059ffd385a1343a83c150
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5664917
Reviewed-by: Maksim Moskvitin <mmoskvitin@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1320857}