[PaymentRequest] Measure source of kCanMakePaymentEnabled user pref
Currently we measure the number of users who do/do not have the
kCanMakePaymentEnabled user pref set. However it is also possible to
measure the source of what changed the pref - e.g. whether it was the
user directly, an extension, an admin policy, etc. This information will
be useful to determine our path forward for the toggle, so this CL
extends the metrics gathering to include that.

To keep the code clean, the metric gathering is also extracted into a
new helper method that can be called by both profile_impl.cc and
payment_request_factory.cc

Bug: 349411309
Change-Id: Id33eaae64e497c962bbdc4a5f6b74b09fee57991
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5655392
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Ryan Sultanem <rsult@google.com>
Reviewed-by: Kavita Soni <kavitasoni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320920}