[PM] NodeInlineData: Migrate ExecutionContextRegistry
This CL gets rid of the ExecutionContextImpl helper as this template
forced the implementation in the header file, but that leads to
circular include now that FrameNodeImpl and WorkerNodeImpl needs
the definition of FrameExecutionContext and WorkerExecutionContext.

Also, the FrameExecutionContext and WorkerExecutionContext is
now created inside FrameNodeImpl::OnJoiningGraph() and
WorkerNodeImpl::OnJoiningGraph() as multiple observers need them
but the order that they are notified is not guaranteed.

Bug: 349425221
Change-Id: Idaeba1998c2ff441e04fe74ac75e3040ddd00b49
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645798
Reviewed-by: Joe Mason <joenotcharles@google.com>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320932}