Fenced frames: Disable recursive prefetch when Fenced Frame untrusted network is disabled
When a prefetch request gets a response header that contains a link
field with preload attribute, it can be changed to a prefetch request
because of the recursive prefetch token.

After fenced frame disables untrusted network, this recursive prefetch
request needs to be blocked. This is done by updating the recursive
prefetch token generator to attach the fenced frame nonce to the
request's IsolationInfo if it is from a fenced frame.

See `change_rel_to_prefetch` in PreloadHelper::LoadLinksFromHeader.

Other changes:
1. Add a test utility function to exempt URLs from network revocation
for fenced frames.
2. Add a test utility function to register a callback that is run
just before PrefetchURLLoader is created and started.

Bug: 1515599
Change-Id: I0bf9ca4988c549f8d1cea71018def93b70145663
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5648422
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Commit-Queue: Xiaochen Zhou <xiaochenzh@chromium.org>
Reviewed-by: Garrett Tanzer <gtanzer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320962}