Fix Supervised User extension install metrics on "Skip Approval"-mode
The metrics for the Extension Installation dialog (the same dialog
that is used by non-supervised users) should be recorded for supervised
user when the "Skip parent approval"-mode is enabled.

This change re-enables observing the states of the
Extension Installation dialog dialog for supervised users,
when parental controls apply on Extensions.
The deprecated state "Request parent approval" is cleaned up
for where is was previously tracked. The state was only
meaningful for an old use-case of the dialog, which no longer exists.
A new state ("Child Accepted") is introduced instead and tracked
when the child proceeds with an extension installation in the
"Skip parent approval"-mode.

Bug: b/321240396, b/349322029
Change-Id: Iadbd2ef4062475be83053536f728db7e85bdc260
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5645669
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Commit-Queue: Anthi Orfanou <anthie@google.com>
Reviewed-by: Courtney Wong <courtneywong@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1320974}