Send salts to VisitedLinkReader for navigations during hashtable build
When the partitioned :visited link hashtable is not building, VisitedLinkReader receives per-origin salts via the NavigationRequest commit parameters. To avoid race conditions, we do not send salts via NavigationRequest when the table is building on another thread. However, per-origin salts are required to hash the triple-partition key representing a link and determine its visited-ness. To ensure each link can access its corresponding salt, this CL updates the VisitedLinkReader with the salts of navigations that took place while the hashtable was building on the DB thread.

Fixed: 349132650
Change-Id: I5802524121c7ec83b64a48aab6b6bc70b93b9f85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5622516
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Commit-Queue: Kyra Seevers <kyraseevers@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Cr-Commit-Position: refs/heads/main@{#1320999}