[Extensions] Add per-extension acknowledgment of MV2 disabled stage
Chrome has announced that MV2 extensions will gradually no longer be
supported in the browser. None of these changes will take effect on any
Chrome channel before June, 2024. See more information at
https://developer.chrome.com/blog/resuming-the-transition-to-mv3.

The first stage of this is a user-visible notice in chrome://extensions,
which can be acknowledged. The second stage will disable unsupported MV2 extensions and will also have a user-visible notice.

This CL adds a per-extension pref acknowledgment of the notice for the
second stage (disable), along with tests for the same. Acknowledging a
notice for a stage is only preserved for such stage. This means if
notice was acknowledged in the first stage (warning), it will not be
acknowledged for the second stage (disable).

Bug: 339061151
Change-Id: Iae488fa78bf723f2c053eed269edc2f5df2bd0d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5664300
Commit-Queue: Emilia Paz <emiliapaz@chromium.org>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321046}