[UB] Create and add feature rows in OnStatusChanged
Previously, the rows (only cookies row currently) were created when CookieControlsContentView was initialized. The label, icon, and other characteristics of the rows would be updated manually and individually in the OnStatusChanged function.

Since we are moving to a framework where we will need to make the feature rows more generalized, this CL gets rid of the creation of the feature row on initialization and updates each row new in the OnStatusChanged function. This will allow us to add new ACT features easily since they will call the same generic AddFeatureRow function enabled in this CL.

We can also use this AddFeatureRow function to deal with managed settings, which will be added in a follow-up CL.

Bug: b:344856056
Change-Id: I383fc39783a20d16f4490aec8497f200d6882831
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5654055
Commit-Queue: Michelle Abreo <michelleabreo@google.com>
Reviewed-by: Fiona Macintosh <fmacintosh@google.com>
Reviewed-by: Kevin Graney <kmg@google.com>
Cr-Commit-Position: refs/heads/main@{#1321047}