Enforce that kFullServerCard is invalid for ShouldShowVirtualCardOption
Full server cards are no longer stored or offered as suggestions in
Autofill prompts - they only exist as a post-fill cached state. As
such, it should not be possible for them to be considered for
showing a virtual card, so enforce that in the code with a
NOTREACHED_NORETURN().

Bug: 40939195
Change-Id: I35028225b96167f2c98d0bea8fec5e17146502c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5660756
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Darwin Yang <darwinyang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1321071}